requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
canvas==0.14.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.0
configparser==7.1.0
contourpy==1.2.1
cookiecutter==2.6.0
cron-converter==1.2.1
cssselect2==0.7.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
Django==5.1.5
django-stubs==5.1.2
django-stubs-ext==5.1.2
django-timezone-utils==0.15.0
docx==0.2.4
env-tools==2.4.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.10.0
fastapi==0.115.8
fastjsonschema==2.19.1
Flask==3.1.0
fonttools==4.53.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
ipykernel==6.27.0
ipython==8.31.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.5.2
json5==0.9.22
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
keyring==25.6.0
kiwisolver==1.4.5
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.8
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.1
matplotlib-inline==0.1.6
matplotlip==0.2
mdurl==0.1.2
mercadopago==2.2.3
mistune==3.0.2
more-itertools==10.6.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
mysql-connector-python==9.5.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.5.8
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.2
openai==1.63.2
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.0.0
plotly==5.23.0
plyer==2.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.41
propcache==0.2.1
protobuf==4.25.5
psutil==5.9.6
psycopg==3.2.4
psycopg-binary==3.2.4
pure-eval==0.2.2
py3Dmol==2.1.0
pycparser==2.21
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.17.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.10.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
RapidFuzz==3.11.0
rdkit==2023.9.5
redis==5.2.1
referencing==0.33.0
regex==2024.11.6
reportlab==4.2.5
requests==2.31.0
requests-toolbelt==1.0.0
RestrictedPython==7.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.5
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
statsd==4.0.1
tenacity==9.0.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tini==4.0.0
tinycss2==1.2.1
tkcalendar==1.6.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
typer==0.15.1
types-python-dateutil==2.9.0.20240315
types-PyYAML==6.0.12.20241230
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.34.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
widgetsnbextension==4.0.10
xlrd==2.0.1
yarl==1.18.3
zstandard==0.23.0
psycopg2-binary
